pyproject.toml
[tool.poetry]
name = "rising-whale"
version = "0.1.0"
description = "Playing around with Streaming and Rising Wave"
authors = ["Franco Bocci <francogbocci@gmail.com>"]
license = "Not open source"
readme = "README.md"
homepage = "https://github.com/privatedumbo/rising-whale"
repository = "https://github.com/privatedumbo/rising-whale"
keywords = ["rising-whale"]
classifiers=[
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/privatedumbo/rising-whale/issues"

[tool.poetry.scripts]
rising-whale = 'rising_whale.cli:app'

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
typer = {extras = ["all"], version = "^0.9.0"}
psycopg2-binary = "^2.9.9"
pyiceberg = {extras = ["sql-postgres"], version = "^0.6.0"}
pydantic-settings = "^2.2.1"
confluent-kafka = "^2.3.0"
dataclasses-avroschema = "^0.57.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.1"
invoke = "^2.2.0"
bump2version = "^1.0.1"
watchdog = {version = "^4.0.0", extras = ["watchmedo"]}
ipdb = "^0.13.13"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.1"
xdoctest = "^1.1.3"
coverage = {version = "^7.4.1", extras = ["toml"]}
pytest-cov = "^4.1.0"

[tool.poetry.group.linters.dependencies]
isort = "^5.13.2"
black = "^24.2.0"
ruff = "^0.2.1"

[tool.poetry.group.typing.dependencies]
mypy = "^1.8.0"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["rising_whale"]

[tool.coverage.report]
fail_under = 100
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug",
  "if settings.DEBUG:",
  "raise AssertionError",
  "raise NotImplementedError",
  "if 0:",
  "if __name__ == __main__:"
]
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
target-version = "py39"
output-format = "full"
line-length = 99
fix = true

[tool.ruff.lint]
select = [
  "E", "F", "W",  # flake8
  "C",  # mccabe
  "I",  # isort
  "N",  # pep8-naming
  # "D",  # flake8-docstrings
  "ANN",  # flake8-annotations
  "S",  # flake8-bandit
  "BLE",  # flake8-blind-except
  "B",  # flake8-bugbear
  "A",  # flake8-builtins
  "G",  # flake8-logging-format
  "ERA",  # eradicate
  "ISC",  # flake8-implicit-str-concat
  "RUF",  # Ruff-specific rules
]
ignore = [
    "ANN101",
    "ANN204",  # Missing return type annotation for special method
]
unfixable = [
  "ERA",  # Don't remove commented-out code
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["rising_whale"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 99
known_third_party = ["invoke", "nox", "nox_poetry"]

[tool.black]
line-length = 99
target-version = ["py39"]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["pytest.*", "invoke.*", "nox.*", "nox_poetry.*"]
allow_redefinition = false
check_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = true
local_partial_types = true
strict_optional = true
strict_equality = true
no_implicit_optional = true
warn_unused_ignores = true
warn_unreachable = true
warn_no_return = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
